turn into CLI
