--- golang/go.mod
@@ -1,16 +1,0 @@
-module github.com/dyne/great-dane
-
-go 1.17
-
-require (
-	github.com/davecgh/go-spew v1.1.1
-	github.com/miekg/dns v1.1.47-0.20220210084238-d521a2cc052f
-)
-
-require (
-	golang.org/x/mod v0.4.2 // indirect
-	golang.org/x/net v0.0.0-20210726213435-c6fcb2dbf985 // indirect
-	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c // indirect
-	golang.org/x/tools v0.1.6-0.20210726203631-07bc1bf47fb2 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-)
